+++ bflibrary/include/bfpixel.h
@@ -29,13 +29,13 @@ extern "C" {
 /** Draw pixel on graphics window, with clipping.
  *
  */
-TbResult LbDrawPixelClip(long x, long y, TbPixel colour);
+TbResult LbDrawPixelClip(s32 x, s32 y, TbPixel colour);
 
 /** Draw pixel on graphics window, without checking coords.
  *
  * This function just sets the pixel to given colour, ignoring DrawFlags.
  */
-TbResult LbDrawPixel(long x, long y, TbPixel colour);
+TbResult LbDrawPixel(s32 x, s32 y, TbPixel colour);
 
 #ifdef __cplusplus
 };
